test: fix tests, add relative to retrieve
